am62x: mcspi: fix PRECISION.LOSS
An implicit cast to a smaller data type can cause a
loss of precision in data.
Modify the type of the variable to fix the issue.

Fixes: SITSW-3820

+++ source/drivers/mcspi/v0/mcspi_v0.c
@@ -2078,7 +2078,7 @@ static void MCSPI_setChDataSize(uint32_t baseAddr, MCSPI_ChObject *chObj,
 
     chObj->csDisable = csDisable;
     /* Calculate buffer access width */
-    chObj->bufWidthShift = MCSPI_getBufWidthShift(dataSize);
+    chObj->bufWidthShift = (uint8_t)MCSPI_getBufWidthShift(dataSize);
 
     /* Calculate data width mask depending on SPI word size */
     chObj->dataWidthBitMask = MCSPI_getDataWidthBitMask(dataSize);
